Task Cleanup/NetworkMode/IAM roles testing for WIndows + TestLaunch scripts
