+++ agent/functional_tests/testdata/taskdefinitions/logdriver-jsonfile-windows/task-definition.json
@@ -1,7 +1,7 @@
 {
   "family": "ecsinteg-json-file-rollover",
   "containerDefinitions": [{
-    "image": "microsoft/windowsservercore:10.0.14393.206",
+    "image": "microsoft/windowsservercore:latest",
     "name": "exit",
     "memory": 10,
     "cpu": 10,
@@ -13,6 +13,6 @@
       }
     },
     "entryPoint": ["powershell"],
-    "command": ["echo", "helloworld"]     
+    "command": ["echo", "helloworld"]
   }]
 }
