+++ agent/functional_tests/testdata/taskdefinitions/network-mode-windows/task-definition.json
@@ -1,0 +1,11 @@
+{
+  "family": "ecsftest-networkmode",
+  "networkMode": "$$$$NETWORK_MODE$$$$",
+  "containerDefinitions": [{
+    "image": "microsoft/windowsservercore:latest",
+    "entryPoint": ["powershell"],
+    "command": ["sleep", "60"],
+    "name": "network-$$$$NETWORK_MODE$$$$",
+    "memory": 50
+  }]
+}
