--- agent/functional_tests/testdata/taskdefinitions/simple-exit-authed-windows/task-definition.json
@@ -1,12 +1,0 @@
-{
-  "family": "ecsftest-simple-exit-authed-windows",
-  "containerDefinitions": [{
-    "image": "microsoft/windowsservercore:10.0.14393.206",
-    "name": "exit",
-    "cpu": 10,
-    "memory": 10,
-    "essential": true,
-    "entryPoint": ["powershell"],
-    "command": ["exit", "42"]
-  }]
-}
