+++ misc/windows-iam/ec2.go
@@ -1,0 +1,40 @@
+// Copyright 2014-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/ec2"
+)
+
+func main() {
+
+	// Set your region for future requests
+
+	svc := ec2.New(session.New(),nil)
+
+	// Call the DescribeInstances Operation
+	regions, err := svc.DescribeRegions(nil)
+	if err != nil {
+		os.Exit(1)
+	}
+	fmt.Println("Success")
+	for _, region := range regions.Regions {
+		fmt.Println(*region.RegionName)
+	}
+	os.Exit(42)
+}
