+++ misc/windows-listen80/listen80.go
@@ -1,0 +1,29 @@
+// +build !windows
+
+// Copyright 2014-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package main
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func pingHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintln(w, "pong")
+}
+func main() {
+	http.HandleFunc("/", pingHandler)
+	http.ListenAndServe("0.0.0.0:80", nil)
+}
